feat: Add action frequency computation and weights
This commit adds the functionality to compute the action frequencies from the Charades_v1_train.csv file and store them in the action_freq.pkl file. It also adds the ability to load the action weights from the action_freq.pkl file. This will be used for weighting the loss function during training.